Remove unrequired cast

+++ libs/ffxiv_parser/src/ex/exh.rs
@@ -20,13 +20,13 @@ impl ExHeader {
         let data: Bytes = package.read_file(&format!("exd/{}.exh", name)).await?;
 
         let header = parse!(data, ExhHeader);
-        let columns = parse!(&data[ExhHeader::SIZE..], header.column_count as usize, ExhColumnDefinition);
+        let columns = parse!(&data[ExhHeader::SIZE..], header.column_count, ExhColumnDefinition);
 
         let pages_base = ExhHeader::SIZE + header.column_count as usize * ExhColumnDefinition::SIZE;
-        let pages = parse!(&data[pages_base..], header.page_count as usize, ExhPage);
+        let pages = parse!(&data[pages_base..], header.page_count, ExhPage);
 
         let languages_base = ExhHeader::SIZE + header.column_count as usize * ExhColumnDefinition::SIZE + header.page_count as usize * ExhPage::SIZE;
-        let languages = parse!(&data[languages_base..], header.language_count as usize, Language);
+        let languages = parse!(&data[languages_base..], header.language_count, Language);
 
         Ok(Self {
             row_size: header.row_size,
